requirements.txt
streamlit
pandas
fastapi
python-dotenv
yfinance
firecrawl-py
fuzzywuzzy
python-Levenshtein
mftool
matplotlib